requirements.txt
pyYaml
nextcord
colorama
requests
mysql-connector-python==8.0.12